refact. clientes
