+++ src/contextos/ventas/cliente/vistas/DetalleCliente/CRMContactos/EdicionCrmContactos.tsx
@@ -1,9 +1,8 @@
-import { QBoton } from "../../../../componentes/atomos/qboton.tsx";
-import { QInput } from "../../../../componentes/atomos/qinput.tsx";
-import { useModelo } from "../../../comun/useModelo.ts";
-import { CrmContacto } from "../diseño.ts";
-import { metaCrmContacto } from "../dominio.ts";
-import "./TabCuentasBanco.css";
+import { QBoton } from "../../../../../../componentes/atomos/qboton.tsx";
+import { QInput } from "../../../../../../componentes/atomos/qinput.tsx";
+import { useModelo } from "../../../../../comun/useModelo.ts";
+import { CrmContacto } from "../../../diseño.ts";
+import { metaCrmContacto } from "../../../dominio.ts";
 
 interface EdicionCrmContactosProps {
   contacto: CrmContacto;
