+++ src/contextos/ventas/factura/vistas/MaestroConDetalleFactura.tsx
@@ -18,16 +18,13 @@ const metaTablaFactura = [
     cabecera: "Código",
   },
   {
-    id: "servido",
-    cabecera: "Servido",
-  },
-  {
     id: "nombre_cliente",
     cabecera: "Cliente",
   },
   {
     id: "total",
     cabecera: "Total",
+    tipo: "moneda",
   },
 ];
